+++ src/traffik/backends/inmemory.py
@@ -101,9 +101,9 @@ class InMemoryBackend(ThrottleBackend[None, HTTPConnectionT]):
     """
     In-memory throttle backend.
 
-    Uses shards improve concurrent access.
+    Uses shards (and hence lock striping) to improve concurrent access.
 
-    Warning: Only use for testing or single-process applications.
+    Warning: Only use for development or single-process applications.
     Does not work across multiple processes/servers.
     """
 
@@ -260,7 +260,9 @@ class InMemoryBackend(ThrottleBackend[None, HTTPConnectionT]):
         """
         if name not in self._named_locks:
             async with self._named_locks_lock:
-                self._named_locks[name] = _AsyncInMemoryLock()
+                lock = _AsyncInMemoryLock()
+                self._named_locks[name] = lock
+                return lock
         return self._named_locks[name]
 
     async def get(